requirement.txt
Django==1.6.5
requests==2.3.0
beautifulsoup4==4.3.2
MySQL-python==1.2.3
Pillow==2.5.3
celery==3.1.15
django-celery==3.1.15
gevent==1.0.1
South==1.0.1
django-redis==3.8.0
